Change tests to conform the prior refactor
