+++ user_management_service/test/unit/commands/vault/auth/AuthSpecFixtures.scala
@@ -1,4 +1,4 @@
-package unit.command.auth
+package unit.commands.vault.auth
 
 import org.scalatest.TestSuite
